infer/requirements.txt
# Core ML Libraries
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
transformers==4.57.1
accelerate==1.10.1
diffusers==0.35.2
tqdm==4.67.1
numpy==2.2.6
pandas==2.2.3
scipy==1.15.3
scikit-learn==1.7.2
scikit-image==0.25.2
pillow==11.2.1
opencv-python==4.12.0.88
matplotlib==3.10.3

# LLM & VLM Specific
vllm==0.8.5.post1
flash-attn==2.7.4.post1
qwen-vl-utils==0.0.11
tiktoken==0.9.0
tokenizers==0.22.1
sentencepiece==0.2.0
protobuf==4.25.7
einops==0.8.1
bitsandbytes==0.48.1
peft==0.17.1

# Tools & Utilities
wandb==0.19.11
hydra-core==1.3.2
pydantic==2.11.4
ray==2.46.0
fastapi==0.115.12
uvicorn==0.34.2
requests==2.32.3
aiohttp==3.11.18
beautifulsoup4==4.13.4
rich==14.0.0
psutil==7.0.0
huggingface-hub==0.36.0
datasets==3.6.0
triton==3.2.0
xformers==0.0.29.post2

# Specialized
# sam2==1.1.0
